Made changes in super user blog part
