+++ src/Components/superuser/blogs/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import AllBlogsPageContainer from '../../drawer/blogs/blogsList';
+
+const SuperuserBlogContainer = () => (
+  <div>
+    <AllBlogsPageContainer
+      isSuperuserRoute
+    />
+  </div>
+);
+
+export default SuperuserBlogContainer;
